move SingleOrMultipleSubtasks to CreateMultipleSubtasksDialog
for cleaner reusability
